requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-genai==1.0.0
Pillow==10.4.0
pdf2image==1.17.0
aiohttp==3.10.0
pydantic-settings==2.6.0
python-multipart==0.0.12
